# docs/swagger.yaml
basePath: /api
definitions:
  ds.Biomarker:
    description: Медицинский биомаркер для исследований
    properties:
      created_at:
        type: string
      description:
        type: string
      full_description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      is_active:
        type: boolean
      max_value:
        type: number
      measure_unit:
        type: string
      min_value:
        type: number
      name:
        type: string
      significance:
        type: number
      updated_at:
        type: string
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@nutriscan.ru
    name: API Support
    url: http://localhost:8081
  description: API для расчета индекса нутритивной недостаточности
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: NutriScan API
  version: "1.0"
paths:
...
  /biomarkers:
    get:
      description: Возвращает список всех биомаркеров с возможностью фильтрации по
        названию
      parameters:
      - description: Фильтр по названию биомаркера
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties: true
            type: object
      summary: Получить список биомаркеров
      tags:
      - Биомаркеры
    post:
      consumes:
      - application/json
      description: Создает новый биомаркер в системе (только для администраторов)
      parameters:
      - description: Данные биомаркера
        in: body
        name: biomarker
        required: true
        schema:
          $ref: '#/definitions/ds.Biomarker'
      produces:
      - application/json
      responses:
        "201":
          description: Биомаркер создан
          schema:
            $ref: '#/definitions/ds.Biomarker'
        "400":
          description: Неверные данные
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Недостаточно прав
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Создать новый биомаркер
      tags:
      - Биомаркеры
  /biomarkers/{id}:
    delete:
      description: Удаляет биомаркер и его изображение из MinIO (только для администраторов)
      parameters:
      - description: ID биомаркера
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Биомаркер удален
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверный ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Недостаточно прав
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Биомаркер не найден
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить биомаркер
      tags:
      - Биомаркеры
    get:
      description: Возвращает детальную информацию о биомаркере
      parameters:
      - description: ID биомаркера
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ
          schema:
            $ref: '#/definitions/ds.Biomarker'
        "400":
          description: Неверный ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Биомаркер не найден
          schema:
            additionalProperties: true
            type: object
      summary: Получить биомаркер по ID
      tags:
      - Биомаркеры
    put:
      consumes:
      - application/json
      description: Обновляет информацию о биомаркере (только для администраторов)
      parameters:
      - description: ID биомаркера
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные биомаркера
        in: body
        name: biomarker
        required: true
        schema:
          $ref: '#/definitions/ds.Biomarker'
      produces:
      - application/json
      responses:
        "200":
          description: Биомаркер обновлен
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверные данные
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Недостаточно прав
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Биомаркер не найден
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Обновить биомаркер
      tags:
      - Биомаркеры
  /biomarkers/{id}/add_to_research:
    post:
      description: Добавляет биомаркер в черновик исследования (только для врачей)
      parameters:
      - description: ID биомаркера
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Биомаркер добавлен в исследование
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверный ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Недостаточно прав
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Биомаркер не найден
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Добавить биомаркер в исследование
      tags:
      - Исследования
  /biomarkers/{id}/image:
    post:
      consumes:
      - multipart/form-data
      description: Загружает изображение для биомаркера в MinIO (только для администраторов)
      parameters:
      - description: ID биомаркера
        in: path
        name: id
        required: true
        type: integer
      - description: Изображение биомаркера
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Изображение загружено
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверные данные
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Недостаточно прав
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Биомаркер не найден
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Загрузить изображение биомаркера
      tags:
      - Биомаркеры
...
securityDefinitions:
  ApiKeyAuth:
    description: Введите "Bearer [ваш_JWT_токен]" для авторизации
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...